Updates from PR review
